requirements.txt
psycopg2
Django
djangorestframework
markdown      
django-filter